feat: added toastr message

+++ app/Http/Controllers/MenuItemCategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\MenuItemCategory;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use Inertia\Inertia;
 use Log;
 
@@ -47,7 +48,7 @@ class MenuItemCategoryController extends Controller
 
         MenuItemCategory::create($valdiated_data);
 
-        return redirect()->route('menu.category');
+        return redirect()->route('menu.category')->with('success', 'Menu Category Created.');
 
 
     }
@@ -74,15 +75,12 @@ class MenuItemCategoryController extends Controller
     public function update(Request $request, MenuItemCategory $menuItemCategory)
     {
         $validated_data = $request->validate([
-            'name' => 'required|unique:menu_item_categories,name'
+            'name' => 'required', Rule::unique('menu_item_categories', 'name')->ignore($menuItemCategory) ,
         ]);
-
-        Log::info($menuItemCategory);
         
          $updated = $menuItemCategory->update($validated_data);
-        Log::info($updated);
 
-        return redirect()->route('menu.category');
+        return redirect()->route('menu.category')->with('success', 'Menu Category Updated.');
     }
 
     /**
@@ -90,8 +88,13 @@ class MenuItemCategoryController extends Controller
      */
     public function destroy(MenuItemCategory $menuItemCategory)
     {
+
+        if($menuItemCategory->menuItems()->count() > 0){
+              return redirect()->back()->with('error', 'Cannot delete menu Category');
+        }
+
          $menuItemCategory->delete();
 
-         return redirect()->route('menu.category');
+         return redirect()->route('menu.category')->with('success', 'Category Deleted Successfully.');
     }
 }
